ai-service/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
python-dotenv
httpx
pdfplumber
python-multipart
